requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-appconf==1.0.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-imagekit==4.0.2
django-modeltranslation==0.16.2
django-rest-framework==0.1.0
djangorestframework==3.12.2
idna==2.10
pilkit==2.0
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
